refactor(desktop): colocate settings modal state into useSettingsModal
Step 3b of the #1043 plan. The Settings modal surface state (open flag,
requested section, provider-catalog sub-open flag) and the three openers
that persist the section to localStorage move into useSettingsModal,
mirroring the use-shell-* seam.

AppShell calls the hook at the original state-declaration site (hook
order preserved) and threads the values to the same consumers: the
chat-model / visual-smoke / bootstrap-subscription hooks that take
openSettings(Section) as deps, the JSX callbacks, and AppShellOverlays.

closeSettings stays inline in AppShell: on close it re-pulls the
onboarding snapshot, the memory-visibility flag, and the default
permission mode - cross-slice orchestration that belongs to the shell,
not the modal. The hook exposes setSettingsOpen /
setSettingsProviderCatalogOpen so closeSettings can clear the surface.

Pure code motion - no behavior change.
