+++ apps/desktop/src/renderer/app-shell.tsx
@@ -4,7 +4,6 @@ import type {
   PermissionMode,
   PlanReminder,
   SessionSummary,
-  SettingsSection,
   ThemePalette,
   ThemePreference,
   UiLocale,
@@ -62,7 +61,6 @@ import { deriveAppShellTurnViewModel } from './app-shell-turn-view-model';
 import { readScrollMotionBehavior } from './scroll-motion-policy';
 import { deriveBranchBanner } from './branch-banner';
 import { applyTheme, applyThemePalette } from './theme';
-import { safeLocalStorageSet } from './browser-storage';
 import { filterSessions, readNavSelection } from './nav-selection';
 import {
   SESSION_LIST_COLLAPSED_WIDTH,
@@ -115,6 +113,7 @@ import { useShellConnections } from './use-shell-connections';
 import { useShellChatModel } from './use-shell-chat-model';
 import { useShellLiveTurn } from './use-shell-live-turn';
 import { useShellLayout } from './use-shell-layout';
+import { useSettingsModal } from './use-settings-modal';
 import {
   isSessionWorkspaceUnavailableError,
   showSessionWorkspaceUnavailableToast,
@@ -213,9 +212,16 @@ export function AppShell({
     refreshConnections,
     handleConnectionEvent,
   } = useShellConnections({ toastApi });
-  const [settingsOpen, setSettingsOpen] = useState(false);
-  const [settingsRequestedSection, setSettingsRequestedSection] = useState<SettingsSection | undefined>(undefined);
-  const [settingsProviderCatalogOpen, setSettingsProviderCatalogOpen] = useState(false);
+  const {
+    settingsOpen,
+    settingsRequestedSection,
+    settingsProviderCatalogOpen,
+    setSettingsOpen,
+    setSettingsProviderCatalogOpen,
+    openSettings,
+    openSettingsSection,
+    openProviderCatalog,
+  } = useSettingsModal();
   const [themePref, setThemePref] = useState<ThemePreference>('auto');
   const [themePalette, setThemePalette] = useState<ThemePalette>('default');
   const [uiLocalePreference, setUiLocalePreference] = useState<UiLocalePreference>('auto');
@@ -1082,13 +1088,8 @@ export function AppShell({
     });
   }
 
-  function openSettings() {
-    setSettingsProviderCatalogOpen(false);
-    setSettingsOpen(true);
-  }
-
   /**
-   * PR-UI-RENDER-2 — single chokepoint for the Markdown internal-URI
+   * PR-UI-RENDER-2 - single chokepoint for the Markdown internal-URI
    * router. Receives a typed `MakaUriDest` from the link override in
    * `<Markdown>` and dispatches to the existing app navigation
    * surfaces:
@@ -1120,26 +1121,6 @@ export function AppShell({
         return _exhaustive;
       }
     }
-  }
-
-  /**
-   * Opens Settings and jumps directly to the named section. Writes the section
-   * to localStorage (so the next cold-open lands there too) and threads it
-   * through `requestedSection` so an already-open Settings modal switches
-   * tabs without close/reopen.
-   */
-  function openSettingsSection(section: SettingsSection) {
-    safeLocalStorageSet('maka-settings-section-v1', section);
-    setSettingsRequestedSection(section);
-    setSettingsProviderCatalogOpen(false);
-    setSettingsOpen(true);
-  }
-
-  function openProviderCatalog() {
-    safeLocalStorageSet('maka-settings-section-v1', 'models');
-    setSettingsRequestedSection('models');
-    setSettingsProviderCatalogOpen(true);
-    setSettingsOpen(true);
   }
 
   function closeSettings() {
